admin css newsletter

+++ action_contact_form.php
@@ -13,13 +13,15 @@ if (isset($_REQUEST['action_type']) && !empty($_REQUEST['action_type'])) {
         $statusMsg = $insert?'Les données ont été insérées.':'Des problèmes sont survenus, reassayez.';
         $_SESSION['statusMsg'] = $statusMsg;
         $referer = isset($_SERVER['HTTP_REFERER']) ? $_SERVER['HTTP_REFERER'] : 'index.php';
-        header('Location: ' . $referer);    }  elseif ($_REQUEST['action_type'] == 'delete') {
+        header('Location: ' . $referer);
+    } elseif ($_REQUEST['action_type'] == 'delete') {
         if (!empty($_GET['id_message_utilisateur'])) {
             $condition = array('id_message_utilisateur' => $_GET['id_message_utilisateur']);
             $delete = $db->delete($tblName, $condition);
             $statusMsg = $delete?'Les données ont été supprimées.':'Des problèmes sont survenus essayez encore.';
             $_SESSION['statusMsg'] = $statusMsg;
             header("Location:admin_messages.php");
+            //var_dump($condition);
         }
     }
 }
